requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
cloudinary==1.41.0
gunicorn==22.0.0
